requirements.txt
streamlit
pandas
requests
beautifulsoup4
spacy
Faker
google-generativeai
lxml[html_clean]